refactor: move interfaces out of `types.ts`
This makes jump-to-definition much more useful.
