+++ src/__tests__/versioning.test.ts
@@ -1,5 +1,7 @@
-import type { IPackageVersionInfo } from '../types'
-import { getClosestMatchingVersion } from '../versioning'
+import {
+  getClosestMatchingVersion,
+  type IPackageVersionInfo,
+} from '../versioning'
 
 describe('getClosestMatchingVersion', () => {
   it('returns the closest matching version', () => {
